resolve-order-issue

+++ models/order_item.go
@@ -13,10 +13,10 @@ type OrderItem struct {
 	Product   Product
 	ProductId uint `gorm:"not null"`
 
-	Slug        string `gorm:"not null"`
-	ProductName string `gorm:"not null"`
-	Price       int    `gorm:"not null"`
-	Quantity    int    `gorm:"not null"`
+	Slug        string  `gorm:"not null"`
+	ProductName string  `gorm:"not null"`
+	Price       float32 `gorm:"not null"`
+	Quantity    int     `gorm:"not null"`
 
 	User   User `gorm:"association_foreignkey:UserId:"`
 	UserId uint `gorm:"default:null"`
